(merge from 3.0)
I think this was tpot's originally:

Fix format types for 64 bit systems.

Andrew Bartlett
(This used to be commit 256b2da7c96e8313f4f98ce700fc7634eaccb72b)

+++ source3/torture/utable.c
@@ -165,8 +165,8 @@ BOOL torture_casetable(int dummy)
 			int c2[MAX_EQUIVALENCE];
 
 			if (size/sizeof(int) >= MAX_EQUIVALENCE) {
-				printf("too many chars match?? size=%d c=0x%04x\n",
-				       size, c);
+				printf("too many chars match?? size=%ld c=0x%04x\n",
+				       (unsigned long)size, c);
 				cli_close(cli, fnum);
 				return False;
 			}
